requirements.txt
alembic==0.8.10
bleach==2.0.0
bokeh==0.12.5
click==6.7
cycler==0.10.0
decorator==4.0.11
dominate==2.3.1
entrypoints==0.2.3
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==19.6.0
html5lib==0.999999999
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Mako==1.0.6
MarkupSafe==0.23
matplotlib==2.1.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
networkx==1.11
notebook==5.0.0
numpy==1.14.0
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
progressbar2==3.34.3
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
pyPdf==1.13
python-dateutil==2.6.0
python-editor==1.0.3
python-utils==2.2.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
runipy==0.1.5
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
SQLAlchemy==1.1.5
tabula-py==0.9.0
terminado==0.6
testpath==0.3.1
tornado==4.5
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.15
widgetsnbextension==2.0.0
WTForms==2.1
xlwt==1.2.0
